# docs/api-spec-docs/openapi.yaml
openapi: 3.0.3
info:
  title: Masjid App API
  description: |
    API for Masjid Application - Prayer times, events, announcements, campaigns, and donations management.
    
    ## Authentication
    - **Admin APIs**: JWT Bearer Token in Authorization header
    - **Member APIs**: API Key in X-API-Key header
    - **Webhooks**: Stripe signature verification
    
    ## Rate Limits
    - Admin APIs: 100 requests/minute per user
    - Member APIs: 60 requests/minute per API key
    - Auth endpoints: 10 requests/minute per IP
  version: 1.0.0
  contact:
    name: Masjid App Support
    email: support@masjid-app.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.{masjid-domain}.com/api/v1
    description: Production server
...
paths:
...
  /admin/auth/login:
    post:
      tags:
        - Admin Auth
      summary: Admin login
      description: Authenticate admin user and receive JWT tokens
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Account locked
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountLockedError'
  /admin/auth/refresh:
    post:
      tags:
        - Admin Auth
      summary: Refresh access token
      description: Get new access token using refresh token cookie
      operationId: adminRefreshToken
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/auth/logout:
    post:
      tags:
        - Admin Auth
      summary: Admin logout
      description: Invalidate refresh token and clear cookies
      operationId: adminLogout
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
  /admin/auth/me:
    get:
      tags:
        - Admin Auth
      summary: Get current user
      description: Get details of currently authenticated admin user
      operationId: getCurrentAdmin
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUserResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /admin/auth/change-password:
    put:
      tags:
        - Admin Auth
      summary: Change password
      description: Change current user's password
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: Password changed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================
  # ADMIN PRAYER TIMES ENDPOINTS
  # ============================================
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for admin authentication

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for mobile app authentication

  # ============================================
  # PARAMETERS
  # ============================================
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Missing or invalid authentication token
            meta:
              timestamp: "2026-01-29T08:00:00Z"
              requestId: req_abc123
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'

  # ============================================
  # SCHEMAS
  # ============================================
  schemas:
...
    Meta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'
    ValidationErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
        meta:
          $ref: '#/components/schemas/Meta'
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
        meta:
          $ref: '#/components/schemas/Meta'

    # Enums
    AdminRole:
      type: string
      enum:
        - super_admin
        - admin
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 8
    LoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
            expiresIn:
              type: integer
              description: Token expiry in seconds
            user:
              $ref: '#/components/schemas/AdminUser'
        meta:
          $ref: '#/components/schemas/Meta'
    RefreshTokenResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            accessToken:
              type: string
            expiresIn:
              type: integer
        meta:
          $ref: '#/components/schemas/Meta'
    AccountLockedError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: ACCOUNT_LOCKED
            message:
              type: string
            lockedUntil:
              type: string
              format: date-time
        meta:
          $ref: '#/components/schemas/Meta'
    ChangePasswordRequest:
      type: object
      required:
        - currentPassword
        - newPassword
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 8

    # Admin User Schemas
    AdminUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
        role:
          $ref: '#/components/schemas/AdminRole'
        isActive:
          type: boolean
        lastLoginAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    AdminUserResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AdminUser'
        meta:
          $ref: '#/components/schemas/Meta'
...